# api/public/swagger-api.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: NLMG API
  description: New Lion HighSchool Sports App API.
host: localhost:8280
basePath: /v1
# basePath: /api
schemes:
  - http
  - https
produces:
  - application/json
securityDefinitions:
  seasonLinkInternal:
    type: oauth2
    description: >
      Exchanges an email/password for an access_token. Only
      to be used by the SeasonLink app.
    # name: access_token
    # in: header
    flow: password
    tokenUrl: http://localhost:8280/v1/login
    scopes: {}
paths:
  /me:
    get:
      summary: Returns the current logged in user.
      operationId: me
      security:
        - seasonLinkInternal: []
      responses:
        200:
          description: User summary
          schema:
            $ref: SuccessEnvelope
            properties:
              meta:
                type: object
              response:
                description: User Model
                $ref: User
        default:
          description: The standard error envelope.
          schema:
            $ref: ErrorEnvelope

  /login:
    post:
      summary: Exchange a username (i.e. email) and password for an access token.
      description: >
        Logging in will return just the access_token in the
        response when successful. Follow up with a call to `/me` if you
        want to get the user details.
      parameters:
        - name: email
          in: formData
          description: >
            The user's email. SeasonLink does not have usernames,
            but the OAuth2 spec requires this to be the param name.
          required: true
          type: string
        - name: password
          in: formData
          description: The user's password.
          required: true
          type: string
      responses:
        200:
          description: Returns the access_token
          schema:
            $ref: SuccessEnvelope
            properties:
              meta:
                type: object
              response:
                properties:
                  access_token:
                    type: string
        default:
          description: The standard error envelope.
          schema:
            $ref: ErrorEnvelope

  /signup:
    post:
      summary: Creates a new user account.
      description: >
        Creating a new user will return an object containing the user
        and the access_token in the response when successful.
      parameters:
        - name: firstName
          in: formData
          description: First name of the user.
          required: false
          type: string
        - name: lastName
          in: formData
          description: Last name of the user.
          required: false
          type: string
        - name: password
          in: formData
          description: The user's password.
          required: true
          type: string
        - name: email
          in: formData
          description: The user's email address.
          required: true
          type: string
      responses:
        200:
          description: Returns the access_token for the new user.
          schema:
            $ref: SuccessEnvelope
            properties:
              meta:
                type: object
              response:
                properties:
                  user:
                    $ref: User
                  access_token:
                    type: string
        default:
          description: The standard error envelope.
          schema:
            $ref: ErrorEnvelope

  /users:
    get:
      description: Returns array of Users.
      summary: Retrieve Users.
      operationId: getUsersList
      responses:
        200:
          description: List all users.
          schema:
            $ref: SuccessEnvelope
            properties:
              meta:
                type: object
              response:
                type: array
                items:
                  $ref: User
        default:
          description: The standard error envelope.
          schema:
            $ref: ErrorEnvelope

  /users/{userId}:
    get:
      parameters:
      - name: userId
        in: path
        description: the Id of the User
        required: true
        type: string
      description: Returns a User based on supplied Id.
      summary: Retrieve a User Object.
      operationId: getUserById
      responses:
        200:
          description:  Return a User.
          schema:
            $ref: SuccessEnvelope
            properties:
              meta:
                type: object
              response:
                $ref: User
        default:
          description: The standard error envelope.
          schema:
            $ref: ErrorEnvelope

  /teams:
    get:
      description: Returns array of Teams.
      summary: Retrieve Teams.
      operationId: getTeamsList
      responses:
        200:
          description: List all Teams.
          schema:
            $ref: SuccessEnvelope
            properties:
              meta:
                type: object
              response:
                type: array
                items:
                  $ref: Team
        default:
          description: The standard error envelope.
          schema:
            $ref: ErrorEnvelope

  /teams/search:
    get:
      parameters:
      - name: q
        in: query
        description: case insensitve text search (Optional)
        required: false
        type: string
      - name: schoolId
        in: query
        description: case sensitve schoolId search (Optional)
        required: false
        type: string
      - name: sport
        in: query
        description: case sensitve sport search (Optional)
        required: false
        type: string
      - name: season
        in: query
        description: case sensitve season search (Optional)
        required: false
        type: string
      - name: level
        in: query
        description: case sensitve level search (Optional)
        required: false
        type: string
      - name: gender
        in: query
        description: case sensitve gender search (Optional)
        required: false
        type: string
      - name: mascot
        in: query
        description: case sensitve mascot search (Optional)
        required: false
        type: string
      description: |
        Returns an array of Teams matching supplied query. Query can be a case insensitve text search and/or a case sensitive key/value pair.
      summary: Retrieve Teams based on search terms or key values
      operationId: getTeamsByQuery
      responses:
        200:
          description: List results
          schema:
            $ref: SuccessEnvelope
            properties:
              meta:
                type: object
              response:
                type: array
                items:
                  $ref: Team

  /teams/follow/{teamId}:
    get:
      parameters:
      - name: teamId
        in: path
        description: the Id of the Team
        required: true
        type: string
      description: Applies teamId to User's followed teams.
      operationId: followTeamById
      responses:
        200:
          description: Return a User.
          schema:
            $ref: SuccessEnvelope
            properties:
              meta:
                type: object
              response:
                $ref: User
        default:
          description: The standard error envelope.
          schema:
            $ref: ErrorEnvelope

  /teams/unfollow/{teamId}:
    get:
      parameters:
      - name: teamId
        in: path
        description: the Id of the Team
        required: true
        type: string
      description: Removes teamId from User's followed teams.
      operationId: unfollowTeamById
      responses:
        200:
          description: Return a User.
          schema:
            $ref: SuccessEnvelope
            properties:
              meta:
                type: object
              response:
                $ref: User
        default:
          description: The standard error envelope.
          schema:
            $ref: ErrorEnvelope

  /teams/{teamId}:
    get:
      parameters:
      - name: teamId
        in: path
        description: the Id of the Team
        required: true
        type: string
      description: Returns a Team based on supplied Id.
      summary: Retrieve a Team Object.
      operationId: getTeamById
      responses:
        200:
          description: Return a Team.
          schema:
            $ref: SuccessEnvelope
            properties:
              meta:
                type: object
              response:
                $ref: Team
        default:
          description: The standard error envelope.
          schema:
            $ref: ErrorEnvelope

  /schools:
    get:
      description: Returns array of Schools.
      summary: Retrieve Schools.
      operationId: getSchoolsList
      responses:
        200:
          description: List all Schools.
          schema:
            $ref: SuccessEnvelope
            properties:
              meta:
                type: object
              response:
                type: array
                items:
                  $ref: School
        default:
          description: The standard error envelope.
          schema:
            $ref: ErrorEnvelope

  /schools/search:
    get:
      parameters:
      - name: q
        in: query
        description: case insensitve text search (Optional)
        required: false
        type: string
      - name: fullName
        in: query
        description: case sensitve fullName search (Optional)
        type: string
      - name: shortName
        in: query
        description: case sensitve shortName search (Optional)
        type: string
      - name: initials
        in: query
        description: case sensitve initials search (Optional)
        type: string
      - name: city
        in: query
        description: case sensitve city search (Optional)
        type: string
      - name: state
        in: query
        description: case sensitve state search (Optional)
        type: string
      - name: zip
        in: query
        description: case sensitve zip search (Optional)
        type: string
      - name: division
        in: query
        description: case sensitve division search (Optional)
        type: string
      - name: mascot
        in: query
        description: case sensitve mascot search (Optional)
        type: string
      - name: logoColorDark
        in: query
        description: case sensitve logoColorDark search (Optional)
        type: string
      - name: logoColorLight
        in: query
        description: case sensitve logoColorLight search (Optional)
        type: string
      - name: storeLink
        in: query
        description: case sensitve storeLink search (Optional)
        type: string
      description: |
        Returns an array of Schools matching supplied query. Query can
        be a case insensitve text search and/or a case sensitive
        key/value pair.
      summary: Retrieve Schools based on search terms or key values
      operationId: getSchoolsByQuery
      responses:
        200:
          description: List results
          schema:
            $ref: SuccessEnvelope
            properties:
              meta:
                type: object
              response:
                type: array
                items:
                  $ref: School

  /schools/{schoolId}:
    get:
      parameters:
      - name: schoolId
        in: path
        description: the Id of the School
        required: true
        type: string
      description: Returns a School based on supplied Id.
      summary: Retrieve a School Object.
      operationId: getSchoolById
      responses:
        200:
          description: Return a School.
          schema:
            $ref: SuccessEnvelope
            properties:
              meta:
                type: object
              response:
                $ref: School
        default:
          description: The standard error envelope.
          schema:
            $ref: ErrorEnvelope

  # /events:

  # /events/search:

  # /events/:eventId:

  # /messages:

  # /messages/search:

  # /messages/:messageId:

definitions:
  ErrorEnvelope:
    description: API error envelope response
    required:
      - meta
    properties:
      meta:
        description: meta data of the error envelope
        type: object
        required:
          - code
          - errorType
          - errorDetail
        properties:
          code:
            type: integer
            format: int32
          errorType:
            description: "Additional information about what went wrong, intended for the developer."
            type: string
          errorDetail:
            description: A localized string intended for the client to display back to the user directly.
            type: string

  SuccessEnvelope:
    description: API success envelope response
    required:
      - meta
      - response
    properties:
      meta:
        description: meta data of the sucess envelope
        type: object
        required:
          - code
        properties:
          code:
            type: integer
            format: int32
      response:
        description: response data of successful API call
        type: object

  User:
    description: User Object
    required:
      - email
      - password
    properties:
      username:
        description: Username of User
        type: string
      firstName:
        description: First name of User
        type: string
      lastName:
        description: Last name of User
        type: string
      email:
        description: Email address of User
        type: string
      password:
        description: Salted passward hash of User
        type: string
      accounts:
        description: List of accounts associated with User
        type: array
        items:
          type: object
      devices:
        description: List of devices associated with User
        type: array
        items:
          type: object
      roles:
        description: List of roles associated with User
        type: array
        items:
          type: object
      createdTime:
        description: Date of User model creation
        type: string
        format: date-time
      lastVisitTime:
        description: Date of User's last visit
        type: string
        format: date-time
      birthday:
        description: Date of User's birthday
        type: string
        format: date-time
      followedTeams:
        description: List of Mongodb ObjectIds of teams followed
        type: array
        items:
          $ref:  ObjectId
      preferences:
        description: List of User preference objects
        type: array
        items:
          type: object

  Team:
    description: Team Object
    required:
      - schoolId
      - sport
      - season
      - level
      - gender
    properties:
      schoolId:
        description: Mongodb ObjectId of School this Team belongs too
        $ref: ObjectId
      sport:
        description: The sport this Team is associated with
        type: string
      season:
        description: The season of this Team's sport
        type: string
      level:
        description: The Team's level of play (Frosh, JV, Varsity)
        type: string
      gender:
        description: The gender of this Team (Male, Female)
        type: string
      mascot:
        description: The Team's mascot
        type: string

  School:
    description: School Object
    required:
      - fullName
      - city
      - state
      - zip
    properties:
      fullName:
        description: Full name of School
        type: string
      shortName:
        description: Short name of School
        type: string
      initials:
        description: Initials of School
        type: string
      city:
        description: City School is located in
        type: string
      state:
        description: State School is located in
        type: string
      zip:
        description: Zip/Postal code of School's location
        type: string
      division:
        description: Division of School
        type: string
      mascot:
        description: Mascot of School
        type: string
      logoColorDark:
        description: Dark color of School's colors
        type: string
      logoColorLight:
        description: Light color of School's colors
        type: string
      storeLink:
        description: Link to School's store
        type: string

  ObjectId:
    description: Mongodb ObjectId definition
    properties:
      id:
        description: Id attribute of ObjectId
        type: string